refactor(capture-sdk): Refactor event package name and structure
PP-438
