+++ capture-sdk/sdk/src/main/java/net/gini/android/capture/camera/CameraFragmentImpl.java
@@ -68,8 +68,8 @@ import net.gini.android.capture.document.ImageMultiPageDocument;
 import net.gini.android.capture.document.QRCodeDocument;
 import net.gini.android.capture.error.ErrorFragment;
 import net.gini.android.capture.error.ErrorType;
-import net.gini.android.capture.events.AnalyticsEvent;
-import net.gini.android.capture.events.AnalyticsScreen;
+import net.gini.android.capture.tracking.user_analytics.AnalyticsEvent;
+import net.gini.android.capture.tracking.user_analytics.AnalyticsScreen;
 import net.gini.android.capture.internal.camera.api.CameraException;
 import net.gini.android.capture.internal.camera.api.CameraInterface;
 import net.gini.android.capture.internal.camera.api.OldCameraController;
