+++ capture-sdk/sdk/src/main/java/net/gini/android/capture/tracking/user_analytics/AnalyticsEventTracker.kt
@@ -2,8 +2,8 @@ package net.gini.android.capture
 
 import android.content.Context
 import com.mixpanel.android.mpmetrics.MixpanelAPI
-import net.gini.android.capture.events.AnalyticsEvent
-import net.gini.android.capture.events.AnalyticsScreen
+import net.gini.android.capture.tracking.user_analytics.AnalyticsEvent
+import net.gini.android.capture.tracking.user_analytics.AnalyticsScreen
 import org.json.JSONObject
 
 interface AnalyticsEventTracker {
