+++ capture-sdk/sdk/src/main/java/net/gini/android/capture/tracking/user_analytics/AnalyticsScreen.kt
@@ -1,4 +1,4 @@
-package net.gini.android.capture.events
+package net.gini.android.capture.tracking.user_analytics
 
 enum class AnalyticsScreen(val screenName: String) {
     CAMERA("camera"),
